requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-extensions==2.2.8
djangorestframework==3.11.0
get==2019.4.13
idna==2.9
post==2019.4.13
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
